requirements.txt
certifi
deepspeed==0.7.3
einops==0.4.1
gpuinfo==1.0.0a7
h5py==2.10.0
ipdb==0.13.9
loralib==0.1.1
matplotlib==3.6.2
munch==2.5.0
Pillow==9.3.0
pretrainedmodels==0.7.4
scikit_learn
tqdm==4.40.0
wandb==0.13.3
wget==3.2
ffmpeg==1.4
ffmpeg-python==0.2.0
ignite==1.1.0
imageio==2.25.1
imageio-ffmpeg==0.4.7
jedi==0.18.1
laion-clap==0.0.6
librosa==0.10.0
moviepy==1.0.3
numba==0.56.4
numpy==1.21.5
opencv-python==4.6.0.66
pandas==1.5.2
protobuf==3.20.1
resampy==0.4.2
ruamel.yaml==0.17.21
scikit-learn==1.1.3
scipy==1.9.3
soundfile==0.12.1
sox==1.4.1
stempeg==0.2.3
tensorboard==2.7.0
tensorboardX==2.1
timm==0.6.12
tokenizers==0.13.2
torch==1.13.0
torchaudio==0.13.0
torchfile==0.1.0
torchlibrosa==0.1.0
torchtext==0.14.0
torchvision==0.14.0
transformers==4.26.1
typer==0.4.2
